moving ingressinalyzer pkg to connlist/internal
